% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annu.pv.df.R
\name{annu.pv.df}
\alias{annu.pv.df}
\title{Calculate present value of annuity}
\usage{
annu.pv.df(pmt,i,n,k)
}
\arguments{
\item{pmt}{:the equal amount of payment of each period}

\item{i}{:interest rate accoding to the period}

\item{n}{:number of periods}

\item{k}{:number of periods deffered}
}
\description{
Calculate present value of an ordinary annuity or an annuity due.
}
\examples{
#annu.pv(100,0.0248,10,4,0)
}

