% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annu.pv.R
\name{annu.pv}
\alias{annu.pv}
\title{Calculate present value of annuity}
\usage{
annu.pv(pmt,i,n,type = 0)
}
\arguments{
\item{pmt}{:the equal amount of payment of each period}

\item{i}{:interest rate accoding to the period}

\item{n}{:number of periods}

\item{type}{:type = 0 for ordinary annuity, type = 1 for annuity due}
}
\description{
Calculate present value of an ordinary annuity or an annuity due.
}
\examples{
#annu.pv(100,0.0248,10,0)
}

