% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaf.R
\name{betaf}
\alias{betaf}
\title{Calculating beta for a company or a select of companies}
\usage{
betaf(x,y,method)
}
\arguments{
\item{x}{:a vector or a data.frame of rate of return of companies}

\item{y}{:name of the independent variable}

\item{method}{:method of calculation; method = 1 for a common expression of beta(see detail); method = 2 using linear regression to estimate the beta}
}
\description{
Calculating beta using common method or linear regression(OLS)
}
\examples{
#betaf(appl,sp500)
}
