% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond.price.R
\name{bond.price}
\alias{bond.price}
\title{Calculate the plain vanilla bond price}
\usage{
bond.price(par,c,n,yield,m)
}
\arguments{
\item{par}{:the face value of the bond}

\item{c}{:the annual coupon rate of the bond}

\item{n}{:number of years}

\item{yield}{:the annual yield to maturity of a bond}

\item{m}{:couponding period in a year}
}
\description{
Calculate the plain vanilla bond price
}
\examples{
#bond.price(1000,0.03,10,0.0248,2)
}
