% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corm.R
\name{corm}
\alias{corm}
\title{Correlation matrix and correlation ranking of a data.frame}
\usage{
corm(x,n)
}
\arguments{
\item{x}{:a data.frame}

\item{n}{:number of decimal points}
}
\description{
Calculating the descriptive statistics of a data.frame and exporting in a data.frame
}
\examples{
#corm(sp1500,3) for correlation matrix of sp1500
}
