% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc}
\alias{desc}
\title{Descriptice statistics of a data.frame}
\usage{
desc(x)
}
\arguments{
\item{x}{:a data.frame}
}
\description{
Calculating the descriptive statistics of a data.frame and exporting in a data.frame
}
\examples{
#desc(sp1500) for descriptive statistics of sp1500
}

