% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmode.R
\name{findmode}
\alias{findmode}
\title{Calculating mode for numeric data}
\usage{
findmode(x)
}
\arguments{
\item{x}{:a numeric variable(vector)}
}
\description{
Calculating mode for numeric data
}
\examples{
# findmode(return)
}

