% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmplot.R
\name{lmplot}
\alias{lmplot}
\title{Diagnostics Graphs for a linear regression}
\usage{
lmplot(x,y)
}
\arguments{
\item{x}{:a linear fitted object}

\item{y}{:the independent variable}
}
\description{
Plotting diagnostic graphs for a fitted object: histogram of the residuals, scatter plot of the fitted values and residuals, scatter plot of fitted values and actual values, time series plot of residuals, and time series plot of actual and fitted values.
}
\examples{
# lmplot(aapl,sp500)
}

