
\name{plotsm}
\alias{plotsm}
\title{Plot scatter smooth plots for a data.frame}
\usage{
plotsm(x,dependent,c)
}
\arguments{
\item{x}{:a dataframe}
\item{dependent}{:the dependent variable}
\item{c}{:is there dummy variable in the data.frame; c = 0 when there is none; c = 1 when there is}
}
\description{
Plotting scatter smooth plots for a data.frame, with name, number and labels.
}
\examples{
# plotsm(JPM-ratios,"price"",0)
}

