% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotts.R
\name{plotts}
\alias{plotts}
\title{Plot time series plots for a data.frame}
\usage{
plotts(x,c)
}
\arguments{
\item{x}{:a dataframe}

\item{c}{:is there dummy variable in the data.frame; c = 0 when there is none; c = 1 when there is}
}
\description{
Plotting time series plots for a data.frame, with name the graphs and number the graphs.
}
\examples{
#plotts(sp500,0)
}

