% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr.R
\name{rr}
\alias{rr}
\title{Calculating rate of return of a vector}
\usage{
rr(x)
}
\arguments{
\item{x}{:a vector of company prices}
}
\description{
Calculating the rate of return of a vector for further analysis, including calculating beta of companies, plotting to see the trend of the stock for technical analysis
}
\examples{
# rr(aapl)
}

