% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploth.R
\name{ploth}
\alias{ploth}
\title{Plot histograms for a data.frame}
\usage{
ploth(x,c)
}
\arguments{
\item{x}{:a dataframe}

\item{c}{:is there dummy variable in the data.frame; c = 0 when there is none; c = 1 when there is}
}
\description{
Plotting histograms for a data.frame. Also the function will name the graphs and number the graphs.
}
\examples{
#ploth(sp500,0) for histograms of sp500 which does not has dummy variables
}

